+++ src/fsharp/service/service.fs
@@ -785,10 +785,13 @@ type BackgroundCompiler(legacyReferenceResolver, projectCacheSize, keepAssemblyC
             cancellable {
                 let! builderOpt, _ = getOrCreateBuilder (ctok, options, userOpName)
                 match builderOpt with
-                | None -> return [||]
+                | None -> return None
                 | Some builder -> 
                     let! checkResults = builder.GetFullCheckResultsAfterFileInProject (ctok, filename)
-                    return checkResults.GetSemanticClassification ctok })
+                    let scopt = checkResults.GetSemanticClassification ctok 
+                    match scopt with
+                    | None -> return None
+                    | Some sc -> return Some (sc.GetView ()) })
 
     /// Try to get recent approximate type check results for a file. 
     member _.TryGetRecentCheckResultsForFile(filename: string, options:FSharpProjectOptions, sourceText: ISourceText option, _userOpName: string) =