+++ vsintegration/src/FSharp.LanguageService/Colorize.fs
@@ -86,7 +86,7 @@ module internal ColorStateLookup_DEPRECATED =
 type internal FSharpScanner_DEPRECATED(makeLineTokenizer : string -> FSharpLineTokenizer) =
     let mutable lineTokenizer = makeLineTokenizer ""
 
-    let mutable extraColorizations : IDictionary<Line0, struct (range * SemanticClassificationType)[] > option = None
+    let mutable extraColorizations : IDictionary<Line0, FSharpSemanticClassificationItem[] > option = None
 
     /// Decode compiler FSharpTokenColorKind into VS TokenColor.
     let lookupTokenColor colorKind =
@@ -147,11 +147,11 @@ type internal FSharpScanner_DEPRECATED(makeLineTokenizer : string -> FSharpLineT
         lineTokenizer <- makeLineTokenizer lineText
 
     /// Adjust the set of extra colorizations and return a sorted list of affected lines.
-    member _.SetExtraColorizations (tokens: struct (range * SemanticClassificationType)[]) =
+    member _.SetExtraColorizations (tokens: FSharpSemanticClassificationItem[]) =
         if tokens.Length = 0 && extraColorizations.IsNone then
             [| |]
         else
-            let newExtraColorizationsKeyed = dict (tokens |> Array.groupBy (fun struct (r, _) -> Line.toZ r.StartLine))
+            let newExtraColorizationsKeyed = dict (tokens |> Array.groupBy (fun item -> Line.toZ item.Range.StartLine))
             let oldExtraColorizationsKeyedOpt = extraColorizations
             extraColorizations <- Some newExtraColorizationsKeyed
             let changedLines =